backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pydantic==2.5.2
python-multipart==0.0.6
firebase-admin==6.2.0
python-dotenv==1.0.0

httpx==0.25.1
numpy<2.0
pandas
shapely
networkx
matplotlib
opencv-python
